PR requested changes

+++ Assets/PatchKit Patcher/Scripts/AppData/FileSystem/DirectoryOperations.cs
@@ -23,8 +23,15 @@ namespace PatchKit.Unity.Patcher.AppData.FileSystem
         /// <exception cref="UnauthorizedAccessException">Unauthorized access.</exception>
         public static bool IsDirectoryEmpty(string dirPath)
         {
-            Assert.IsFalse(string.IsNullOrEmpty(dirPath));
-            Assert.IsTrue(Directory.Exists(dirPath));
+            if (string.IsNullOrEmpty(dirPath))
+            {
+                throw new ArgumentException("Value cannot be null or empty", "dirPath");
+            }
+
+            if (!Directory.Exists(dirPath))
+            {
+                throw new ArgumentException("Directory must exist", "dirPath");
+            }
 
             try
             {
@@ -51,10 +58,18 @@ namespace PatchKit.Unity.Patcher.AppData.FileSystem
         /// <param name="path">The path.</param>
         /// <exception cref="ArgumentException"><paramref name="path"/> is null or empty.</exception>
         /// <exception cref="UnauthorizedAccessException">Unauthorized access.</exception>
-        public static void CreateParentDirectory(string path)
+        public static void CreateParentDirectory(string path, CancellationToken cancellationToken)
         {
-            Assert.IsFalse(string.IsNullOrEmpty(path));
+            if (string.IsNullOrEmpty(path))
+            {
+                throw new ArgumentException("Value cannot be null or empty", "path");
+            }
 
+            RetryStrategy.TryExecute(() => CreateParentDirectoryInternal(path, cancellationToken), cancellationToken);
+        }
+
+        private static void CreateParentDirectoryInternal(string path, CancellationToken cancellationToken)
+        {
             try
             {
                 DebugLogger.Log(string.Format("Creating parent directory for <{0}>.", path));
@@ -63,7 +78,7 @@ namespace PatchKit.Unity.Patcher.AppData.FileSystem
 
                 if (!string.IsNullOrEmpty(dirPath))
                 {
-                    CreateDirectory(dirPath);
+                    CreateDirectory(dirPath, cancellationToken);
                 }
 
                 DebugLogger.Log("Parent directory created.");
@@ -81,10 +96,18 @@ namespace PatchKit.Unity.Patcher.AppData.FileSystem
         /// <param name="dirPath">The directory path.</param>
         /// <exception cref="ArgumentException"><paramref name="dirPath"/> is null or empty.</exception>
         /// <exception cref="UnauthorizedAccessException">Unauthorized access.</exception>
-        public static void CreateDirectory(string dirPath)
+        public static void CreateDirectory(string dirPath, CancellationToken cancellationToken)
         {
-            Assert.IsFalse(string.IsNullOrEmpty(dirPath));
+            if (string.IsNullOrEmpty(dirPath))
+            {
+                throw new ArgumentException("Value cannot be null or empty", "dirPath");
+            }
 
+            RetryStrategy.TryExecute(() => CreateDirectoryInternal(dirPath), cancellationToken);
+        }
+
+        private static void CreateDirectoryInternal(string dirPath)
+        {
             try
             {
                 DebugLogger.Log(string.Format("Creating directory <{0}>.", dirPath));
@@ -110,17 +133,21 @@ namespace PatchKit.Unity.Patcher.AppData.FileSystem
         /// <exception cref="UnauthorizedAccessException">Unauthorized access.</exception>
         public static void Delete(string dirPath, CancellationToken cancellationToken, bool recursive = false)
         {
-            Assert.IsFalse(string.IsNullOrEmpty(dirPath));
-            Assert.IsTrue(Directory.Exists(dirPath));
+            if (string.IsNullOrEmpty(dirPath))
+            {
+                throw new ArgumentException("Value cannot be null or empty", "dirPath)");
+            }
+
+            if (!Directory.Exists(dirPath))
+            {
+                throw new ArgumentException("Directory must exist", "dirPath");
+            }
 
             RetryStrategy.TryExecute(() => DeleteInternal(dirPath, recursive), cancellationToken);
         }
 
         private static void DeleteInternal(string dirPath, bool recursive)
         {
-            Assert.IsFalse(string.IsNullOrEmpty(dirPath));
-            Assert.IsTrue(Directory.Exists(dirPath));
-
             try
             {
                 DebugLogger.Log(string.Format("Deleting directory {0}<{1}>.",
